LPS-171715 search-experiences-web: Move language functions into a directory in utils
