+++ modules/dxp/apps/search-experiences/search-experiences-web/src/main/resources/META-INF/resources/sxp_blueprint_admin/js/edit_sxp_blueprint/EditSXPBlueprintForm.js
@@ -36,7 +36,8 @@ import {addParams, fetchData, fetchPreviewSearch} from '../utils/fetch';
 import filterAndSortClassNames from '../utils/functions/filter_and_sort_class_names';
 import getResultsError from '../utils/functions/get_results_error';
 import isDefined from '../utils/functions/is_defined';
-import {formatLocaleWithUnderscores, renameKeys} from '../utils/language';
+import formatLocaleWithUnderscores from '../utils/language/format_locale_with_underscores';
+import renameKeys from '../utils/language/rename_keys';
 import {setStorageAddSXPElementSidebar} from '../utils/sessionStorage';
 import cleanUIConfiguration from '../utils/sxp_element/clean_ui_configuration';
 import getUIConfigurationValues from '../utils/sxp_element/get_ui_configuration_values';
