+++ modules/dxp/apps/search-experiences/search-experiences-web/src/main/resources/META-INF/resources/sxp_blueprint_admin/js/edit_sxp_blueprint/add_sxp_element_sidebar/index.js
@@ -36,7 +36,7 @@ import {
 	DEFAULT_SXP_ELEMENT_ICON,
 } from '../../utils/data';
 import {addParams, fetchData} from '../../utils/fetch';
-import {getLocalizedText} from '../../utils/language';
+import getLocalizedText from '../../utils/language/get_localized_text';
 import {setStorageAddSXPElementSidebar} from '../../utils/sessionStorage';
 import isElementInactiveFromNonCompanyIndex from '../../utils/sxp_element/is_element_inactive_from_noncompany_index';
