+++ modules/dxp/apps/search-experiences/search-experiences-web/src/main/resources/META-INF/resources/sxp_blueprint_admin/js/edit_sxp_blueprint/clause_contributors_sidebar/ManagementToolbar.js
@@ -20,7 +20,7 @@ import {ManagementToolbar as FrontendManagementToolbar} from 'frontend-js-compon
 import React, {useState} from 'react';
 
 import {ALL, ASCENDING} from '../../utils/constants';
-import {sub} from '../../utils/language';
+import sub from '../../utils/language/sub';
 
 function ManagementToolbar({
 	allItems,
