+++ modules/dxp/apps/search-experiences/search-experiences-web/src/main/resources/META-INF/resources/sxp_blueprint_admin/js/edit_sxp_blueprint/query_builder_tab/SelectTypes.js
@@ -19,7 +19,7 @@ import {ManagementToolbar} from 'frontend-js-components-web';
 import React, {useContext, useState} from 'react';
 
 import ThemeContext from '../../shared/ThemeContext';
-import {sub} from '../../utils/language';
+import sub from '../../utils/language/sub';
 
 function SelectTypes({
 	onFrameworkConfigChange,
