+++ modules/dxp/apps/search-experiences/search-experiences-web/src/main/resources/META-INF/resources/sxp_blueprint_admin/js/shared/LearnMessage.js
@@ -13,7 +13,7 @@ import ClayLink from '@clayui/link';
 import getCN from 'classnames';
 import React, {useContext} from 'react';
 
-import {getLocalizedLearnMessageObject} from '../utils/language';
+import getLocalizedLearnMessageObject from '../utils/language/get_localized_learn_message_object';
 import ThemeContext from './ThemeContext';
 
 /**
