+++ modules/dxp/apps/search-experiences/search-experiences-web/src/main/resources/META-INF/resources/sxp_blueprint_admin/js/shared/sxp_element/index.js
@@ -23,13 +23,13 @@ import React, {useContext, useEffect, useState} from 'react';
 import {ASSET_CATEGORY_ID} from '../../utils/constants';
 import {DEFAULT_SXP_ELEMENT_ICON} from '../../utils/data';
 import isDefined from '../../utils/functions/is_defined';
+import getLocalizedText from '../../utils/language/get_localized_text';
 import cleanUIConfiguration from '../../utils/sxp_element/clean_ui_configuration';
 import getSXPElementJSON from '../../utils/sxp_element/get_sxp_element_json';
 import isElementInactiveFromNonCompanyIndex from '../../utils/sxp_element/is_element_inactive_from_noncompany_index';
 import {INPUT_TYPES} from '../../utils/types/inputTypes';
 import {PreviewModalWithCopyDownload} from '../PreviewModal';
 import ThemeContext from '../ThemeContext';
-import {getLocalizedText} from './../../utils/language';
 import DateInput from './DateInput';
 import FieldInput from './FieldInput';
 import FieldListInput from './FieldListInput';
