+++ modules/dxp/apps/search-experiences/search-experiences-web/src/main/resources/META-INF/resources/sxp_blueprint_admin/js/utils/sxp_element/get_sxp_element_json.js
@@ -9,7 +9,8 @@
  * distribution rights of the Software.
  */
 
-import {formatLocaleWithUnderscores, renameKeys} from '../language';
+import formatLocaleWithUnderscores from '../language/format_locale_with_underscores';
+import renameKeys from '../language/rename_keys';
 import replaceTemplateVariable from './replace_template_variable';
 
 /**
