+++ modules/dxp/apps/search-experiences/search-experiences-web/src/main/resources/META-INF/resources/sxp_blueprint_admin/js/utils/validation.js
@@ -13,7 +13,7 @@ import {ASSET_CATEGORY_ID} from './constants';
 import {ERROR_MESSAGES} from './errorMessages';
 import isDefined from './functions/is_defined';
 import isEmpty from './functions/is_empty';
-import {sub} from './language';
+import sub from './language/sub';
 import {INPUT_TYPES} from './types/inputTypes';
 
 export function validateBoost(configValue, type) {
